changed DecimalFormat format method to use double instead of long.

+++ Usage/src/main/java/io/pivotal/tola/cfapi/usage/model/AUsage.java
@@ -22,7 +22,7 @@ public class AUsage {
     public String getTotalGbPerAis(){
 
         DecimalFormat df = new DecimalFormat("#.###");
-        return df.format(this.getTotalMbPerAis() / 1024);
+        return df.format( (double) this.getTotalMbPerAis() / 1024);
 
     }
